DRP interleaver implementation
Includes code for implementing DRP interleavers of different sizes and read-write vector combinations. Not very efficient coding as it uses case statement, which is interpreted as logic by the FPGA. A more efficient way might be to store the interleaver combinations as LUTs.
IMPORTANT: Keith's students are looking at different interleavers, so DRP may not be used finally.